+++ plugin/src/main/java/com/denizenscript/denizen/events/world/RaidFinishesScriptEvent.java
@@ -1,8 +1,6 @@
 package com.denizenscript.denizen.events.world;
 
-import com.denizenscript.denizen.events.BukkitScriptEvent;
 import com.denizenscript.denizen.objects.PlayerTag;
-import com.denizenscript.denizen.utilities.world.RaidData;
 import com.denizenscript.denizencore.objects.ObjectTag;
 import com.denizenscript.denizencore.objects.core.ListTag;
 import org.bukkit.entity.Player;
@@ -10,20 +8,11 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.raid.RaidFinishEvent;
 
-public class RaidFinishesScriptEvent extends BukkitScriptEvent implements Listener {
+public class RaidFinishesScriptEvent extends RaidScriptEvent<RaidFinishEvent> implements Listener {
 
     public RaidFinishesScriptEvent() {
+        super(true);
         registerCouldMatcher("raid finishes");
-    }
-
-    public RaidFinishEvent event;
-
-    @Override
-    public boolean matches(ScriptPath path) {
-        if (!runInCheck(path, event.getRaid().getLocation())) {
-            return false;
-        }
-        return super.matches(path);
     }
 
     @Override
@@ -35,8 +24,6 @@ public class RaidFinishesScriptEvent extends BukkitScriptEvent implements Listen
                     list.addObject(new PlayerTag(player));
                 }
                 return list;
-            case "raid":
-                return RaidData.toMap(event.getRaid());
         }
         return super.getContext(name);
     }
