+++ plugin/src/main/java/com/denizenscript/denizen/events/world/RaidSpawnsWaveScriptEvent.java
@@ -1,8 +1,6 @@
 package com.denizenscript.denizen.events.world;
 
-import com.denizenscript.denizen.events.BukkitScriptEvent;
 import com.denizenscript.denizen.objects.EntityTag;
-import com.denizenscript.denizen.utilities.world.RaidData;
 import com.denizenscript.denizencore.objects.ObjectTag;
 import com.denizenscript.denizencore.objects.core.ListTag;
 import org.bukkit.entity.Raider;
@@ -10,20 +8,11 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.raid.RaidSpawnWaveEvent;
 
-public class RaidSpawnsWaveScriptEvent extends BukkitScriptEvent implements Listener {
+public class RaidSpawnsWaveScriptEvent extends RaidScriptEvent<RaidSpawnWaveEvent> implements Listener {
 
     public RaidSpawnsWaveScriptEvent() {
+        super(true);
         registerCouldMatcher("raid spawns wave");
-    }
-
-    public RaidSpawnWaveEvent event;
-
-    @Override
-    public boolean matches(ScriptPath path) {
-        if (!runInCheck(path, event.getRaid().getLocation())) {
-            return false;
-        }
-        return super.matches(path);
     }
 
     @Override
@@ -37,8 +26,6 @@ public class RaidSpawnsWaveScriptEvent extends BukkitScriptEvent implements List
                     raiders.addObject(new EntityTag(raider));
                 }
                 return raiders;
-            case "raid":
-                return RaidData.toMap(event.getRaid());
         }
         return super.getContext(name);
     }
