+++ plugin/src/main/java/com/denizenscript/denizen/events/world/RaidStopsScriptEvent.java
@@ -1,7 +1,5 @@
 package com.denizenscript.denizen.events.world;
 
-import com.denizenscript.denizen.events.BukkitScriptEvent;
-import com.denizenscript.denizen.utilities.world.RaidData;
 import com.denizenscript.denizencore.objects.ObjectTag;
 import com.denizenscript.denizencore.objects.core.ElementTag;
 import com.denizenscript.denizencore.utilities.CoreUtilities;
@@ -9,21 +7,17 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.raid.RaidStopEvent;
 
-public class RaidStopsScriptEvent extends BukkitScriptEvent implements Listener {
+public class RaidStopsScriptEvent extends RaidScriptEvent<RaidStopEvent> implements Listener {
 
     public RaidStopsScriptEvent() {
+        super(true);
         registerCouldMatcher("raid stops");
         registerSwitches("reason");
     }
 
-    public RaidStopEvent event;
-
     @Override
     public boolean matches(ScriptPath path) {
-        if (path.switches.containsKey("reason") && !path.checkSwitch("reason", CoreUtilities.toLowerCase(event.getReason().name()))) {
-            return false;
-        }
-        if (!runInCheck(path, event.getRaid().getLocation())) {
+        if (!runGenericSwitchCheck(path, "reason", event.getReason().name())) {
             return false;
         }
         return super.matches(path);
@@ -34,8 +28,6 @@ public class RaidStopsScriptEvent extends BukkitScriptEvent implements Listener 
         switch (name) {
             case "reason":
                 return new ElementTag(CoreUtilities.toLowerCase(event.getReason().name()), true);
-            case "raid":
-                return RaidData.toMap(event.getRaid());
         }
         return super.getContext(name);
     }
